+++ lib/image_ffmpeg.dart
@@ -8,6 +8,7 @@ export 'src/ffmpeg.dart' show Ffmpeg;
 export 'src/web_config.dart' show FfmpegWeb;
 export 'src/models.dart'
     show
+        BoxAverageAlphaMode,
         EncodedImage,
         FfmpegCapabilities,
         FfmpegException,